Driver support for true quantization in eager mode  (#20)
* Driver supports proper quantization now

* Provide llama3 example

* initial support for i4 quantization

* Add test for explicit quantization

* Refactor bindings

* Fix linux compilation error

* Faster tests

* Update to OV 2024.1

* Add a bunch of examples

* fix issues with scikit-learn

* fix

* Add test for int4

* Refactor quantization code to accomodate for int4

* Add llava example

+++ intel_npu_acceleration_library/nn/linear.py
@@ -127,7 +127,7 @@ class QuantizedLinear(torch.nn.Module):
             raise RuntimeError(
                 f"Quantized weight must be in torch.int8 dtype instead of {self.weight.dtype}"
             )
-        self.scale = scale.to(torch.float32)
+        self.scale = scale
         self.outC, self.inC = self.weight.shape
         self.bias = bias
         self.op_id = str(uuid.uuid4())
